Split helpers to cargo_loader
